Remove swc_plugin_shared
