+++ plugin/Cargo.toml
@@ -11,6 +11,7 @@ version = "0.0.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+abi_stable = "0.9.3"
 anyhow = "1.0.41"
 serde = "1.0.126"
 serde_json = "1.0.64"
@@ -18,4 +19,3 @@ swc_atoms = {path = "../atoms"}
 swc_common = {path = "../common"}
 swc_ecma_ast = {path = "../ecmascript/ast"}
 swc_ecma_visit = {path = "../ecmascript/visit"}
-swc_plugin_shared = {path = "./shared"}
