+++ plugin/runner/Cargo.toml
@@ -19,7 +19,6 @@ swc_atoms = {path = "../../atoms/"}
 swc_common = {path = "../../common/"}
 swc_ecma_ast = {path = "../../ecmascript/ast/"}
 swc_ecma_parser = {path = "../../ecmascript/parser/"}
-swc_plugin_shared = {path = "../shared"}
 
 [dev-dependencies]
 swc_ecma_codegen = {path = "../../ecmascript/codegen/"}
