--- plugin/shared/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-authors = ["강동윤 <kdy1997.dev@gmail.com>"]
-description = "Base for swc plugins"
-documentation = "https://rustdoc.swc.rs/swc_plugin/"
-edition = "2018"
-license = "Apache-2.0/MIT"
-name = "swc_plugin_shared"
-repository = "https://github.com/swc-project/swc.git"
-version = "0.0.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-abi_stable = "0.9.3"
